ls-tree: introduce function "fast_path()"
The generic "show_tree_fmt()" is slower than "show_tree()", so
we want to take the fast path if possible.

when "--format=<format>" is passed, "fast_path()" will determine
whether to use "show_tree()" or insist on using "show_tree_fmt()"
by a try of finding out if the built-int format is hit.

This commit take out the related codes from "cmd_ls_tree()" and
package them into a new funtion "fast_path()".

Explain it a little bit further, whether fast_path is hit or not,
the final correctness should not break. Abstracting a separate method
helps improve the readability of "cmd_ls_tree()" and the cohesiveness
and extensibility of fast path logic.

Signed-off-by: Teng Long <dyroneteng@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>